`[Pull request to display  active tab which has been selected]`

https://dev.azure.com/MULABS/Bank%20Projects/_workitems/edit/481/
**Type of change**

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)

**How Has This Been Tested?**

What tests did you run to verify your changes? Provide instructions so we can reproduce. Please also list any relevant
details for your test configuration

- [x] Test A
![image.png](https://dev.azure.com/MULABS/f1790893-6e4c-43a4-8d41-fc1f3b360635/_apis/git/repositories/ec482b3c-c4c3-400a-99de-40c52e2276ca/pullRequests/262/attachments/image.png)
- [ ] Test B

**Check And Confirm All The Option**

#### PR Push And Review Checklist:

- [x] I don't have any commented out code and blocks
- [x] I have formatted my code and have no formating issue. (use plugins)
- [x] I have checked and confirmed, am only pushing my changes . No changes by mistake or unknown changes
- [x] I have no unknown changes as part of this PR

##### Standards Checklist:

- [x] This review follows the Meraki and Labs Code Review Principal
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation. Updated documents are linked below:
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes

Related work items: #481

+++ src/app/statement/(protected)/settings/close-account/close.account.tsx
@@ -24,8 +24,10 @@ const CloseAccount = ({ onModalCancel }: CloseAccountProps) => {
 
     try {
       await closeAccount(userId);
-      sessionStorage.clear();
-      localStorage.clear();
+      if (typeof window !== "undefined") {
+        sessionStorage.clear();
+        localStorage.clear();
+      }
       notification.info({
         message:
           "Account Closure in Progress. The Account Will Be Closed Permanently After 30 Days of Inactivity.",
